// src/utils/interfaces/nutrients.ts
export interface ISingleNutrients {
  food_name: string;
  nf_calories: number;
  nf_cholesterol: number;
  nf_dietary_fiber: number;
  nf_potassium: number;
  nf_protein: number;
  nf_saturated_fat: number;
  nf_sodium: number;
  nf_sugars: number;
  nf_total_carbohydrate: number;
  nf_total_fat: number;
  photo: {
    thumb: string;
  };
}